+++ packages/data-types/src/lib/interfaces/crafting-unlock-by-mastery.interface.ts
@@ -1,8 +1,0 @@
-import { MinimalItem } from "@ci/data-types";
-
-export interface CraftingUnlockByMastery {
-    masteryLevel: number,
-    masteryType: string,
-    key: string;
-    item?: MinimalItem
-}