+++ packages/pak-files-parser/src/app/cooking-recipe-unlocked-by-mastery-db.generator.ts
@@ -1,0 +1,11 @@
+import { readAsset } from '../util/functions';
+import { RawUnlockByMastery } from "../interfaces/raw-unlock-by-mastery.interface";
+import { Datatable } from "../interfaces/datatable.interface";
+import { UnlockedByMasteryDbGenerator } from "./unlocked-by-mastery-db.generator";
+
+export class CookingRecipeUnlockedByMasteryDbGenerator extends UnlockedByMasteryDbGenerator {
+
+    // ProjectCoral Content ProjectCoral Data Cooking
+    recipesDB: Datatable<RawUnlockByMastery>[] = readAsset<Datatable<RawUnlockByMastery>[]>('DT_CookingRecipeUnlockByMastery.json');
+
+}