+++ packages/pak-files-parser/src/interfaces/raw-unlock-by-mastery.interface.ts
@@ -1,0 +1,25 @@
+export interface RawUnlockByMastery {
+    "masteryLevel": number,
+    "desc": string,
+    "unlockRecipe": {
+        [masteryNameEnum: string]: {
+            "craftingList": [
+                {
+                    "useCustomID": boolean,
+                    "item": {
+                        "data": {
+                            "DataTable": {
+                                "ObjectName": string,
+                                "ObjectPath": string
+                            },
+                            "RowName": string
+                        },
+                        "itemID": string
+                    },
+                    "customID": string
+                }
+            ]
+        }
+    }
+
+}